mobot/v0.6/requirements.txt
openai
sse_starlette
tavily-python
gradio
pyyaml
pinecone-client
tiktoken
openai
langchain
networkx
numexpr
langchainhub
langchain_experimental
pymysql
bs4